# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Warehouse API",
        "description": "API for managing warehouses",
        "version": "1.0"
    },
    "paths": {
...
        "/stocks": {
            "get": {
                "operationId": "55288330091eb5e3826918f6e24e168f",
                "responses": {
                    "200": {
                        "description": "List of stocks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "city_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "Плеханова, дом 2"
                                            },
                                            "lat": {
                                                "type": "number",
                                                "example": 59.836997
                                            },
                                            "lng": {
                                                "type": "number",
                                                "example": 30.494354
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-12-16T18:36:51Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-12-16T18:36:51Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "operationId": "c4c8faac67f0090446d70e3a484dbde9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "city_id": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "maximum": "90",
                                        "minimum": "-90"
                                    },
                                    "lng": {
                                        "type": "number",
                                        "maximum": "180",
                                        "minimum": "-180"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stock created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "city_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "Плеханова, дом 2"
                                        },
                                        "lat": {
                                            "type": "number",
                                            "example": 59.836997
                                        },
                                        "lng": {
                                            "type": "number",
                                            "example": 30.494354
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-16T18:36:51Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-16T18:36:51Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stocks/{id}": {
            "get": {
                "operationId": "e9d95630afecafd1614c8debeeed7d4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stock details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "city_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "Плеханова, дом 2"
                                        },
                                        "lat": {
                                            "type": "number",
                                            "example": 59.836997
                                        },
                                        "lng": {
                                            "type": "number",
                                            "example": 30.494354
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-16T18:36:51Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-16T18:36:51Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stock/{id}": {
            "put": {
                "operationId": "2956bc25ac4e6cabb3132198ca9e09b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "city_id": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "maximum": "90",
                                        "minimum": "-90"
                                    },
                                    "lng": {
                                        "type": "number",
                                        "maximum": "180",
                                        "minimum": "-180"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "city_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "Плеханова, дом 2"
                                        },
                                        "lat": {
                                            "type": "number",
                                            "example": 59.836997
                                        },
                                        "lng": {
                                            "type": "number",
                                            "example": 30.494354
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-16T18:36:51Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-16T18:36:51Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "operationId": "b788b8bcd7930ef770ba7e0c5a049488",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Stock deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stock/nearby": {
            "post": {
                "operationId": "2a81286ea3e20ef13321a0f34fb42b0f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lat": {
                                        "type": "number"
                                    },
                                    "lng": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nearest stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stock": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "city_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Плеханова, дом 2"
                                                },
                                                "lat": {
                                                    "type": "number",
                                                    "example": 59.836997
                                                },
                                                "lng": {
                                                    "type": "number",
                                                    "example": 30.494354
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-16T18:36:51Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-16T18:36:51Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "distance": {
                                            "type": "number",
                                            "example": 1.234
                                        },
                                        "city": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Minsk"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-16T18:36:51Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-16T18:36:51Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}